requirements.txt
fastapi
uvicorn
websockets
wsproto
clbb-hermes
geopandas
pandas
pyarrow
requests